Update patch set 2
Patch Set 2:

Dear Kanagaraj, 

The amsterdam branch was created from label v1.1.0 (which is what was released for cli). I noticed that at this point only the version.properties has the correct 1.1.0 version reflected but the pom files do not. The scripts we use to bump the version took care of bumping the patch version which is what is expected but if the pom files were having the incorrect versions from the start, we might need to do these modifications manually. 

Do you want me to manually reflect the v1.1.1 versions on these files?

Patch-set: 2
Label: Code-Review=0