Update patch set 12
Patch Set 12: Code-Review+1

> (2 comments)

The cloud configuration can be specified during service creation and it used to be required. I'm not positive but I think it is still required for VCPE using VNF-API. VID does not use the full functionality of ONAP components as it was meant to demo ONAP, so using the output as a guide is not comprehensive. The expectation was that most ONAP users would have OSS/BSS software that would make the REST calls directly, but things like CLI and ONAP External API are meant to supplement or replace VID/OSS/BSS direct interaction. 

If you can create using integration use-cases it's probably fine to leave cloudConfig out. Ideally we would remove the json boiler plate for cloud config if these params are not specified rather than remove the ability to use them as SO does use these values in some instances. However from my knowledge of CLI, I don't know if this is trivial.

https://gerrit.onap.org/r/gitweb?p=so.git;a=blob;f=common/src/main/java/org/onap/so/serviceinstancebeans/Service.java;h=c368f67907ed8016d6f4885075f89c4d93a0674b;hb=refs/heads/dublin#l40 

https://gerrit.onap.org/r/gitweb?p=so.git;a=blob;f=common/src/main/java/org/onap/so/serviceinstancebeans/CloudConfiguration.java;h=4a284b55ec152313dea8fa539649261a8724ca56;hb=refs/heads/dublin

https://gerrit.onap.org/r/gitweb?p=so.git;a=blob;f=mso-api-handlers/mso-api-handler-infra/src/main/java/org/onap/so/apihandlerinfra/validation/CloudConfigurationValidation.java;h=2207c52048d2bbd061ffdbc9a1671fd23c880a6e;hb=refs/heads/dublin#l76

Patch-set: 12
Reviewer: Gerrit User 667 <667@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1